requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
PyYAML==5.4.1
requests==2.25.1
smmap==3.0.5
urllib3==1.26.3
